refactor: splite quiz Practic section in technicalInterview component to smaller and extract custom hooks
